+++ src/augment-pr-entry/augment-pr-entry.js
@@ -2,59 +2,7 @@
 // @jsx h
 
 import { h } from 'dom-chef'
-import { ago } from 'time-ago'
-import { getRepoURL } from '../page-detect'
-import api, { type PullRequest, type PullRequestActivity } from '../api'
-
-import './augment-pr-entry.css'
-import linkifyTargetBranch from '../linkify-target-branch/linkify-target-branch'
-
-const buildSourceBranchNode = (branchName: string): HTMLElement => {
-    const repoUrl = getRepoURL()
-    return (
-        <span class="__rbb-pull-request-source-branch">
-            <span class="ref-label">
-                <span class="ref branch">
-                    <span class="name" aria-label={`branch ${branchName}`}>
-                        <a
-                            style={{ color: '#707070' }}
-                            title={branchName}
-                            href={`https://bitbucket.org/${repoUrl}/branch/${branchName}`}
-                        >
-                            {branchName}
-                        </a>
-                    </span>
-                </span>
-            </span>
-        </span>
-    )
-}
-
-export const addSourceBranch = (prNode: HTMLElement, prData: PullRequest) => {
-    const sourceBranchNode = buildSourceBranchNode(prData.source.branch.name)
-    const arrow = prNode.querySelector('span.aui-iconfont-devtools-arrow-right')
-    if (arrow && arrow.parentElement) {
-        arrow.parentElement.insertBefore(sourceBranchNode, arrow)
-    }
-}
-
-export const addCreationDate = (prNode: HTMLElement, prData: PullRequest) => {
-    const date = new Date(prData.created_on)
-    const dateString = date.toDateString()
-    const creationDateNode = (
-        <div title={dateString} datetime={prData.created_on}>
-            Created on {dateString} ({ago(date)})
-        </div>
-    )
-
-    const prNumberAndTimestamp = prNode.querySelector(
-        '.pr-number-and-timestamp'
-    )
-    if (prNumberAndTimestamp) {
-        prNumberAndTimestamp.append(<br />)
-        prNumberAndTimestamp.appendChild(creationDateNode)
-    }
-}
+import api, { type PullRequestActivity } from '../api'
 
 export const addUsernameWithLatestUpdate = (
     prNode: HTMLElement,
@@ -89,16 +37,11 @@ export const addUsernameWithLatestUpdate = (
 }
 
 export default function augmentPrEntry(prNode: HTMLElement) {
-    linkifyTargetBranch(prNode)
-
     const prId = prNode.dataset.pullRequestId
 
-    api.getPullrequest(prId).then(prData => {
-        if (prData) {
-            addSourceBranch(prNode, prData)
-            addCreationDate(prNode, prData)
-        }
-    })
+    if (!prId) {
+        return
+    }
 
     api.getPullrequestActivity(prId).then(prActivity => {
         if (prActivity) {